Only check ttfb on finalUrl
