+++ lighthouse-core/audits/time-to-first-byte.js
@@ -6,8 +6,8 @@
 'use strict';
 
 const Audit = require('./audit');
-const Util = require('../report/v2/renderer/util');
 const URL = require('../lib/url-shim');
+const Util = require('../report/v2/renderer/util');
 
 const TTFB_THRESHOLD = 200;
 const TTFB_THRESHOLD_BUFFER = 15;
@@ -24,7 +24,7 @@ class TTFBMetric extends Audit {
       informative: true,
       helpText: 'Time To First Byte identifies the time at which your server sends a response.' +
         '[Learn more](https://developers.google.com/web/tools/chrome-devtools/network-performance/issues).',
-      requiredArtifacts: ['devtoolsLogs']
+      requiredArtifacts: ['devtoolsLogs', 'URL']
     };
   }
 
@@ -41,47 +41,26 @@ class TTFBMetric extends Audit {
   static audit(artifacts) {
     const devtoolsLogs = artifacts.devtoolsLogs[Audit.DEFAULT_PASS];
 
-    return Promise.all([
-      artifacts.requestNetworkRecords(devtoolsLogs),
-      artifacts.requestCriticalRequests(devtoolsLogs)
-    ])
-      .then(([networkRecords, criticalRequests]) => {
-        const results = [];
+    return artifacts.requestNetworkRecords(devtoolsLogs)
+      .then((networkRecords) => {
         let displayValue;
 
-        criticalRequests.forEach(request => {
-          const networkRecord = networkRecords.find(record => record._requestId === request.id);
+        const finalUrl = artifacts.URL.finalUrl;
+        const thresholdDisplay = Util.formatMilliseconds(TTFB_THRESHOLD, 1);
+        const finalUrlRequest = networkRecords.find(record => record._url === finalUrl);
+        const ttfb = TTFBMetric.caclulateTTFB(finalUrlRequest);
+        const passed = ttfb < TTFB_THRESHOLD + TTFB_THRESHOLD_BUFFER;
 
-          if (networkRecord && networkRecord._timing) {
-            const ttfb = TTFBMetric.caclulateTTFB(networkRecord);
-            results.push({
-              url: URL.getURLDisplayName(networkRecord._url),
-              ttfb: Util.formatMilliseconds(Math.round(ttfb), 1),
-              rawTTFB: ttfb
-            });
-          }
-        });
-
-        const recordsOverBudget = results.filter(row =>
-            row.rawTTFB > TTFB_THRESHOLD + TTFB_THRESHOLD_BUFFER
-        );
-
-        if (recordsOverBudget.length) {
-          const thresholdDisplay = Util.formatMiliseconds(TTFB_THRESHOLD, 1);
-          const recordsOverBudgetDisplay = Util.formatNumber(recordsOverBudget.length);
-          displayValue = `${recordsOverBudgetDisplay} critical request(s) went over` +
+        if (passed) {
+          displayValue = `${URL.getURLDisplayName(finalUrl)} is below` +
+            ` the ${thresholdDisplay} threshold`;
+        } else {
+          displayValue = `${URL.getURLDisplayName(finalUrl)} went over` +
             ` the ${thresholdDisplay} threshold`;
         }
 
-        const headings = [
-          {key: 'url', itemType: 'url', text: 'URL'},
-          {key: 'ttfb', itemType: 'text', text: 'Time To First Byte (ms)'},
-        ];
-
         return {
-          rawValue: recordsOverBudget.length === 0,
-          results,
-          headings,
+          rawValue: passed,
           displayValue,
         };
       });
