+++ lighthouse-core/gather/computed/critical-request-chains.js
@@ -6,6 +6,7 @@
 'use strict';
 
 const ComputedArtifact = require('./computed-artifact');
+const WebInspector = require('../../lib/web-inspector');
 
 class CriticalRequestChains extends ComputedArtifact {
 
@@ -13,18 +14,115 @@ class CriticalRequestChains extends ComputedArtifact {
     return 'CriticalRequestChains';
   }
 
-  generateChain(request) {
-    return {
-      request: {
-        id: request.id,
-        url: request.url,
+  /**
+   * For now, we use network priorities as a proxy for "render-blocking"/critical-ness.
+   * It's imperfect, but there is not a higher-fidelity signal available yet.
+   * @see https://docs.google.com/document/d/1bCDuq9H1ih9iNjgzyAL0gpwNFiEP4TZS-YLRp_RuMlc
+   * @param  {any} request
+   */
+  static isCritical(request) {
+    const resourceTypeCategory = request._resourceType && request._resourceType._category;
+
+    // XHRs are fetched at High priority, but we exclude them, as they are unlikely to be critical
+    // Images are also non-critical.
+    // Treat any images missed by category, primarily favicons, as non-critical resources
+    const nonCriticalResourceTypes = [
+      WebInspector.resourceTypes.Image._category,
+      WebInspector.resourceTypes.XHR._category
+    ];
+    if (nonCriticalResourceTypes.includes(resourceTypeCategory) ||
+        request.mimeType && request.mimeType.startsWith('image/')) {
+      return false;
+    }
+
+    return ['VeryHigh', 'High', 'Medium'].includes(request.priority());
+  }
+
+  static extractChain(networkRecords) {
+    networkRecords = networkRecords.filter(req => req.finished);
+
+    // Build a map of requestID -> Node.
+    const requestIdToRequests = new Map();
+    for (const request of networkRecords) {
+      requestIdToRequests.set(request.requestId, request);
+    }
+
+    // Get all the critical requests.
+    /** @type {!Array<NetworkRequest>} */
+    const criticalRequests = networkRecords.filter(CriticalRequestChains.isCritical);
+
+    const flattenRequest = request => {
+      return {
+        url: request._url,
         startTime: request.startTime,
         endTime: request.endTime,
         responseReceivedTime: request.responseReceivedTime,
-        transferSize: request.transferSize,
-      },
-      children: {},
+        transferSize: request.transferSize
+      };
     };
+
+    // Create a tree of critical requests.
+    const criticalRequestChains = {};
+    for (const request of criticalRequests) {
+      // Work back from this request up to the root. If by some weird quirk we are giving request D
+      // here, which has ancestors C, B and A (where A is the root), we will build array [C, B, A]
+      // during this phase.
+      const ancestors = [];
+      let ancestorRequest = request.initiatorRequest();
+      let node = criticalRequestChains;
+      while (ancestorRequest) {
+        const ancestorIsCritical = CriticalRequestChains.isCritical(ancestorRequest);
+
+        // If the parent request isn't a high priority request it won't be in the
+        // requestIdToRequests map, and so we can break the chain here. We should also
+        // break it if we've seen this request before because this is some kind of circular
+        // reference, and that's bad.
+        if (!ancestorIsCritical || ancestors.includes(ancestorRequest.requestId)) {
+          // Set the ancestors to an empty array and unset node so that we don't add
+          // the request in to the tree.
+          ancestors.length = 0;
+          node = undefined;
+          break;
+        }
+        ancestors.push(ancestorRequest.requestId);
+        ancestorRequest = ancestorRequest.initiatorRequest();
+      }
+
+      // With the above array we can work from back to front, i.e. A, B, C, and during this process
+      // we can build out the tree for any nodes that have yet to be created.
+      let ancestor = ancestors.pop();
+      while (ancestor) {
+        const parentRequest = requestIdToRequests.get(ancestor);
+        const parentRequestId = parentRequest.requestId;
+        if (!node[parentRequestId]) {
+          node[parentRequestId] = {
+            request: flattenRequest(parentRequest),
+            children: {}
+          };
+        }
+
+        // Step to the next iteration.
+        ancestor = ancestors.pop();
+        node = node[parentRequestId].children;
+      }
+
+      if (!node) {
+        continue;
+      }
+
+      // If the node already exists, bail.
+      if (node[request.requestId]) {
+        continue;
+      }
+
+      // node should now point to the immediate parent for this request.
+      node[request.requestId] = {
+        request: flattenRequest(request),
+        children: {}
+      };
+    }
+
+    return criticalRequestChains;
   }
 
   /**
@@ -33,35 +131,8 @@ class CriticalRequestChains extends ComputedArtifact {
    * @return {!Promise<!Object>}
    */
   compute_(devtoolsLog, artifacts) {
-    return artifacts.requestCriticalRequests(devtoolsLog)
-      .then(criticalRequests => {
-        // Create a tree of critical requests.
-        const criticalRequestChains = {};
-        const mappedRequests = {};
-
-        let request = criticalRequests.shift();
-        while(request) {
-          if (!mappedRequests[request.id]) {
-            mappedRequests[request.id] = this.generateChain(request);
-          }
-
-          const node = mappedRequests[request.id];
-          const parent = request.parent;
-          if (parent) {
-            if (!mappedRequests[parent.id]) {
-              mappedRequests[parent.id] = this.generateChain(parent);
-            }
-
-            mappedRequests[parent.id].children[request.id] = node;
-          } else {
-            criticalRequestChains[request.id] = node;
-          }
-
-          request = criticalRequests.shift();
-        }
-
-        return criticalRequestChains;
-      });
+    return artifacts.requestNetworkRecords(devtoolsLog)
+      .then(CriticalRequestChains.extractChain);
   }
 }
