+++ lighthouse-core/test/audits/time-to-first-byte-test.js
@@ -10,66 +10,33 @@ const assert = require('assert');
 
 /* eslint-env mocha */
 describe('Performance: time-to-first-byte audit', () => {
-  it.skip('fails when ttfb is higher than 215ms', () => {
+  it('fails when ttfb of root document is higher than 215ms', () => {
     const networkRecords = [
-      {_url: 'https://google.com/', _requestId: '0', _timing: {receiveHeadersEnd: 500, sendEnd: 200}},
-      {_url: 'https://google.com/styles.css', _requestId: '1', _timing: {receiveHeadersEnd: 414, sendEnd: 200}},
+      {_url: 'https://example.com/', _requestId: '0', _timing: {receiveHeadersEnd: 500, sendEnd: 200}},
+      {_url: 'https://google.com/styles.css', _requestId: '1', _timing: {receiveHeadersEnd: 450, sendEnd: 200}},
       {_url: 'https://google.com/image.jpg', _requestId: '2', _timing: {receiveHeadersEnd: 600, sendEnd: 400}},
     ];
-    const artifacts = {networkRecords: {defaultPass: networkRecords}};
-
-    artifacts.requestCriticalRequestChains = () => {
-      return Promise.resolve(
-        {
-          '0': {
-            request: {url: networkRecords[0]._url},
-            children: {
-              '1': {
-                request: {url: networkRecords[1]._url},
-                children: {},
-              }
-            },
-          },
-          '1': {
-            request: {url: networkRecords[2]._url},
-            children: {},
-          }
-        }
-      );
+    const artifacts = {
+      devtoolsLogs: {[TimeToFirstByte.DEFAULT_PASS]: []},
+      requestNetworkRecords: () => Promise.resolve(networkRecords),
+      URL: {finalUrl: 'https://example.com/'},
     };
 
     return TimeToFirstByte.audit(artifacts).then(result => {
       assert.strictEqual(result.rawValue, false);
-      assert.ok(result.displayValue.includes('1 request(s)'));
     });
   });
 
-  it.skip('succeeds when no request is under 215ms', () => {
+  it('succeeds when ttfb of root document is lower than 215ms', () => {
     const networkRecords = [
-      {_url: 'https://google.com/', _requestId: '0', _timing: {receiveHeadersEnd: 300, sendEnd: 200}},
-      {_url: 'https://google.com/styles.css', _requestId: '1', _timing: {receiveHeadersEnd: 414, sendEnd: 200}},
+      {_url: 'https://example.com/', _requestId: '0', _timing: {receiveHeadersEnd: 300, sendEnd: 200}},
+      {_url: 'https://google.com/styles.css', _requestId: '1', _timing: {receiveHeadersEnd: 450, sendEnd: 200}},
       {_url: 'https://google.com/image.jpg', _requestId: '2', _timing: {receiveHeadersEnd: 600, sendEnd: 400}},
     ];
-    const artifacts = {networkRecords: {defaultPass: networkRecords}};
-
-    artifacts.requestCriticalRequestChains = () => {
-      return Promise.resolve(
-        {
-          '0': {
-            request: {url: networkRecords[0]._url},
-            children: {
-              '1': {
-                request: {url: networkRecords[1]._url},
-                children: {},
-              }
-            },
-          },
-          '1': {
-            request: {url: networkRecords[2]._url},
-            children: {},
-          }
-        }
-      );
+    const artifacts = {
+      devtoolsLogs: {[TimeToFirstByte.DEFAULT_PASS]: []},
+      requestNetworkRecords: () => Promise.resolve(networkRecords),
+      URL: {finalUrl: 'https://example.com/'},
     };
 
     return TimeToFirstByte.audit(artifacts).then(result => {
